Extract DpgNodeBuilder; rename NodePaletteWidget to DpgNodeListBuilder
- New DpgNodeBuilder (src/ui/dpg_node_builder.py) encapsulates all DPG
  widget construction for a single node: file dialog, per-param rows,
  and input/output ports. Param rendering is dispatched via a
  NodeParamType -> method dict, so adding a new param type is a one-line
  registration plus a new _build_*_param method -- NodeEditorPage does
  not need to change.
- NodeEditorPage now delegates to self._node_builder.build(node); the
  param/port/dialog helpers and associated _SAVE_EXTS/_OPEN_EXTS
  constants have moved out of the page module.
- Renamed node_palette_widget.py -> dpg_node_list_builder.py and the
  class NodePaletteWidget -> DpgNodeListBuilder (PAYLOAD_TYPE value
  likewise "NODE_PALETTE" -> "NODE_LIST").
