+++ src/ui/dpg_node_list_builder.py
@@ -3,25 +3,26 @@ from __future__ import annotations
 import dearpygui.dearpygui as dpg
 
 from core.node_registry import NodeEntry, NodeRegistry
+from ui._types import DpgTag
 
 
-class NodePaletteWidget:
-    """Drag-source palette listing all registered nodes by category.
+class DpgNodeListBuilder:
+    """Drag-source list widget enumerating registered nodes by category.
 
     The widget owns its container, search input and per-entry drag payloads.
     Consumers drop nodes onto a canvas by setting
-    ``payload_type=NodePaletteWidget.PAYLOAD_TYPE`` on their drop target; the
+    ``payload_type=DpgNodeListBuilder.PAYLOAD_TYPE`` on their drop target; the
     ``drag_data`` delivered to the drop callback is the :class:`NodeEntry`.
 
     Must be instantiated inside an active DPG container context.
     """
 
-    PAYLOAD_TYPE: str = "NODE_PALETTE"
+    PAYLOAD_TYPE: str = "NODE_LIST"
 
     _CATEGORIES: tuple[str, ...] = ("Sources", "Filters", "Sinks")
 
     def __init__(self, registry: NodeRegistry, *, width: int = 200) -> None:
-        self._items: list[tuple[int | str, str]] = []
+        self._items: list[tuple[DpgTag, str]] = []
 
         with dpg.child_window(width=width, height=-1, border=True):
             dpg.add_input_text(hint="Search…", width=-1, callback=self._on_search)
@@ -45,7 +46,7 @@ class NodePaletteWidget:
             dpg.add_text(f"+ {entry.display_name}")
         self._items.append((tag, entry.display_name.lower()))
 
-    def _on_search(self, sender: int | str, value: str) -> None:
+    def _on_search(self, sender: DpgTag, value: str) -> None:
         query = value.strip().lower()
         for tag, text in self._items:
             dpg.configure_item(tag, show=(not query or query in text))
